Add userUtilities
