+++ backend/src/services/IAccountManager.ts
@@ -9,4 +9,5 @@ export interface UserData {
     userName: string;
     firstName: string;
     lastName: string;
+    isAdmin?: boolean;
 }
